mqtt-elasticsearch-bridge/requirements.txt
paho-mqtt
elasticsearch>=6.0.0,<7.0.0
 